reorg protocol & header methods
