+++ src/fs/protocol/index.ts
@@ -1,3 +1,3 @@
-export * from './basic'
-export * from './header'
+export * as basic from './basic'
+export * as pub from './public'
 export * as priv from './private'
