+++ src/fs/v1/PrivateTree.ts
@@ -5,7 +5,7 @@ import * as keystore from '../../keystore'
 import * as protocol from '../protocol'
 import * as check from '../protocol/private/types/check'
 import * as pathUtil from '../path'
-import * as header from './header'
+import * as metadata from '../metadata'
 import * as namefilter from '../protocol/private/namefilter'
 import { isObject, Maybe } from '../../common'
 import { NonEmptyPath, SyncHookDetailed, UnixTree } from '../types'
@@ -40,7 +40,7 @@ export default class PrivateTree implements UnixTree {
     const fs = new PrivateTree(mmpt, baseName, baseKey)
     await fs.addNode({
       metadata: {
-        ...header.emptyMetadata(),
+        ...metadata.empty(),
         name: 'private'
       },
       bareNameFilter,
@@ -105,7 +105,7 @@ export default class PrivateTree implements UnixTree {
     } else if (!check.isPrivateFile(file)) {
       throw new Error('Can not `cat` a directory')
     }
-    return protocol.getPrivateFile(file.content, file.key)
+    return protocol.basic.getEncryptedFile(file.content, file.key)
   }
 
   async exists(path: string): Promise<boolean> {
@@ -119,7 +119,7 @@ export default class PrivateTree implements UnixTree {
       throw new Error("Can not change a directory to a file")
     }
     const ownKey = file === null ? await keystore.genKeyStr() : file.key
-    const { cid } = await protocol.putPrivateFile(content, ownKey)
+    const { cid } = await protocol.basic.putEncryptedFile(content, ownKey)
     return this.addWithFn(path, async (parent, name, key) => {
       if(file !== null) {
         return protocol.priv.updateFile(file as PrivateFile, cid)
@@ -130,7 +130,7 @@ export default class PrivateTree implements UnixTree {
   }
 
   async addNode(node: DecryptedNode, key: string): Promise<NamedAddResult> {
-    const { cid, size } = await protocol.putPrivateFile(node, key)
+    const { cid, size } = await protocol.basic.putEncryptedFile(node, key)
     const filter = await namefilter.addRevision(node.bareNameFilter, key, node.revision)
     const name = await namefilter.toPrivateName(filter)
     await this.mmpt.add(name, cid)
